Optimized query for report page.
Now, we go backwards. For a total amount of information, we don't
care where it comes from, only that it's accurate. We do this by getting all
Sites with SiteDetail entries, then grabbing the data we want from those.

This has been tested agains the previous method, and returns the same results.
We can conclude this is an accurate representation of the data from that.

refs #14745

Conflicts:

	orvsd_central/views.py